Modified all files except for readme.md, style.css and index.html
