+++ tabaCountriesHolidays/holidays/fetchHolidays.js
@@ -1,6 +1,7 @@
 import { getResHolidays } from "./getResHolidays.js";
 
-export async function fetchHolidays(getDataHolidays, countrySelect, yearSelect, holidaysList, tableHolidays) {
+export async function fetchHolidays(params) {
+    const {getDataHolidays, countrySelect, yearSelect, holidaysList, tableHolidays} = params;
     const getCountry = countrySelect.value;
     const getYear = yearSelect.value;
 
@@ -11,7 +12,8 @@ export async function fetchHolidays(getDataHolidays, countrySelect, yearSelect, 
             holidaysList.push(...holidays.response.holidays);
             getResHolidays(holidaysList, tableHolidays);
         } catch (error) {
+            console.error("Помилка при отриманні списку свят:", error);
             throw new Error("Помилка при отриманні списку свят:", error); 
         }
     } 
-}
+};
